% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpectedParticipantTable.R
\name{ExpectedParticipantTable}
\alias{ExpectedParticipantTable}
\title{ExpectedParticipantTable}
\usage{
ExpectedParticipantTable(
  recruitment_profile,
  end_timepoint_duration = NULL,
  ...
)
}
\arguments{
\item{recruitment_profile}{An data.frame The number of the simulation.}

\item{...}{Not used at present.}
}
\value{
A data.frame. Contains the following columns:
\itemize{
\item{site}
\item{timepoint}
\item{participants}
}
}
\description{
A function that uses the manifest file and the downloaded Virtual
Population data from KerusCloud to create a table with the expected number
of participants per site and at the end of each supplied time period.
}
\author{
Gareth Burns
}
