% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractRecruitmentProfile.R
\name{ExtractRecruitmentProfile}
\alias{ExtractRecruitmentProfile}
\title{ExtractRecruitmentProfile}
\usage{
ExtractRecruitmentProfile(recruitment_json, ...)
}
\arguments{
\item{recruitment_json}{An list. The number of the simulation.}

\item{...}{Not used at present.}
}
\value{
A data.frame. Contains the following columns:
\itemize{
\item{site}
\item{timepoint}
\item{participants}
}
}
\description{
A function that uses the metadata file and the downloaded Virtual
Population data from KerusCloud to extract the recruitment profile and return
in a format for functional programming.
}
\author{
Gareth Burns
}
