% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractTimepoints.R
\name{ExtractTimepoints}
\alias{ExtractTimepoints}
\title{Select KerusCloud Virtual Population Simulations File}
\usage{
ExtractTimepoints(
  metadata,
  variables = NULL,
  infer_baseline = FALSE,
  as_list = FALSE,
  ...
)
}
\arguments{
\item{metadata}{An list. The KerusCloud Virtual Population metadata.}

\item{variables}{A character vector. Manifest file downloaded from KerusCloud.}

\item{infer_baseline}{A logical value. Infer missing values as timepoint = 0.}

\item{as_list}{A logical value. See details. Defaults to FALSE.}

\item{...}{Not used at present.}
}
\value{
A data.frame containing the following columns:
\itemize{
\item{Variable}
\item(Timepoint)
}
simulation file from the simulation and scenarios specified.
}
\description{
A function that uses the Metadata JSON to extract the time points
of each variable and create a long \code{data.frame}.
}
\author{
Gareth Burns
}
