% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractVariableMetadata.R
\name{ExtractVariableMetadata}
\alias{ExtractVariableMetadata}
\title{Read KerusCloud Virtual Population Simulations}
\usage{
ExtractVariableMetadata(
  metadata,
  variable_names = NULL,
  variable_types = NULL,
  ...
)
}
\arguments{
\item{metadata}{\strong{A list} Virtual Population file name(s).}

\item{variable_names}{\strong{A character vector}. Name(s) of the}

\item{...}{Not used at present.}

\item{variarable_types}{\strong{A character value}. The names of the variable
types. Can be one of:
\itemize{
\item{"Exponential"},
\item{"Log-Normal"},
\item{"Normal"},
\item{"Uniform"},
\item{"Weibull"},
\item{"Binomial"},
\item{"Negative-Binomial"},
\item{"Multinomial"},
\item{"Group"},
\item{"Recruitment"},
\item{"Time-to-Event"},
\item{"Derived Multinomial"},
\item{"Operational Derivation"},
\item{"Continious RM"},
\item{"Irreversible Event"},
\item{"Reversible Event"}
}}
}
\value{
A list containing the metadata of the selected variables from
variable_names and variable_list arguments.
}
\description{
A function optimised to read in KerusCloud Virtual Population
simulations.
}
\details{
The meta_data can also be the directory of the
}
\author{
Gareth Burns
}
