% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadKerusSimulations.R
\name{ReadKerusSimulations}
\alias{ReadKerusSimulations}
\title{Read KerusCloud Virtual Population Simulations}
\usage{
ReadKerusSimulations(
  files,
  directory = getwd(),
  manifest_file = NULL,
  as_dataframe = TRUE,
  ...
)
}
\arguments{
\item{files}{\strong{A character vector} Virtual Population file name(s).}

\item{directory}{\strong{A character value} Directory of the Virtual Population
files. Defaults to current working directory.}

\item{manifest_file}{\strong{A data.frame} manifest file output by KerusCloud
Virtual Population download. Optional.}

\item{as_dataframe}{\strong{A logical value.} Defaults to \code{TRUE}. If \code{TRUE} a
\code{data.frame} is returned, if \code{FALSE} a list is returned.}

\item{...}{Named arguments that can be passed to read.csv}
}
\value{
A list of simulations scenarios or the total number of simulation
scenarios in the manifest file.
}
\description{
A function optimised to read in KerusCloud Virtual Population
simulations.
}
\details{
If the manifest file is supplied it will retrieve the respective simulation
and scenario details and append it as a column.
\emph{Note}: If you supply the \code{nrows} and \code{colClasses} arguments to \code{...}
this will make the import of files quicker.
}
\author{
Gareth Burns
}
