% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectVirtualPopulation.R
\name{SelectVirtualPopulation}
\alias{SelectVirtualPopulation}
\title{Select KerusCloud Virtual Population Simulations File}
\usage{
SelectVirtualPopulation(
  simulation,
  manifest_file,
  selected_scenarios = NULL,
  ...
)
}
\arguments{
\item{simulation}{An integer. The number of the simulation.}

\item{manifest_file}{A data.file. Manifest file downloaded from KerusCloud.}

\item{selected_scenarios}{A named list. Name corresponds to the scenario type
and the element corresponds to the desired label of the scenario.}

\item{...}{Not used at present.}
}
\value{
A character value. Corresponds to the individual
simulation file from the simulation and scenarios specified.
}
\description{
A function that uses the manifest file and the downloaded Virtual
Population data from KerusCloud to identify simulation files.
}
\author{
Gareth Burns
}
